% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrad.R
\name{extrad}
\alias{extrad}
\title{monthly data from extraterrestrial radiation}
\usage{
extrad(x, blaney.criddle)
}
\arguments{
\item{x}{latitude of study zone}

\item{blaney.criddle}{a list of 3 dataframes: first list is light hours per day,
second list is maximum hours and third list is extraterrestial radiation.}
}
\description{
function for generate monthly data from extraterrestrial radiation by latitude.
}
