% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdwgen.R
\name{pdwgen}
\alias{pdwgen}
\title{values extraction of PISCO daily files}
\usage{
pdwgen(x)
}
\arguments{
\item{x}{a dataframe with PISCO file names (in netCDF format) of precipitation
minimum and maximum temperature, longitude and latitude of station.}
}
\description{
function for extract weather data from PISCO (SENAMHI, Peru) daily files from
precipitation, minimum and maximum temperature. PISCO is Peruvian Interpolated
Data of the Senamhi’s Climatological and Hydrologycal Observations.
}
