% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swrad.R
\name{swrad}
\alias{swrad}
\title{daily data from short wave radiation}
\usage{
swrad(x, blaney.criddle)
}
\arguments{
\item{x}{latitude of study zone}

\item{blaney.criddle}{a list of 3 dataframes: first list is light hours per day,
second list is maximum hours and third list is extraterrestial radiation.}
}
\description{
function for generate daily values from short wave radiation by latitude.
}
