% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmwgen.R
\name{pmwgen}
\alias{pmwgen}
\title{values average of PISCO daily dataframe}
\usage{
pmwgen(x)
}
\arguments{
\item{x}{a dataframe with PISCO daily dataframe from precipitation, minimum
and maximum temperature.}
}
\description{
function for calculate values average of PISCO daily dataframe from precipitation,
minimum and maximum temperature.
}
